Change lovescore calculation
I have modified the method for calculating the distance between points in space. Previously, the distance was calculated using the sum of the absolute differences between the coordinates of the points, each multiplied by their respective coefficients (Manhattan distance). Now, the distance is calculated using a weighted Euclidean distance, which also accounts for specific coefficients for each axis.